pyproject.toml
[tool.poetry]
name = "exchange"
package-mode = false
version = "0.1.0"
description = ""
authors = ["Ivan Krasnopolsky <i1.krasnopolskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
celery = {extras = ["redis"], version = "^5.4.0"}
certifi = "^2024.8.30"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"
django = "^5.0.7"
django-cors-headers = "^4.4.0"
django-environ = "^0.11.2"
django-redis = "^5.4.0"
pydantic = "^2.9.0"
websockets = "^13.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
profile = "black"
